+++ src/solver/mod.rs
@@ -3,4 +3,4 @@ mod roux;
 mod search;
 pub use cfop::cfop;
 pub use roux::roux;
-pub use search::search;
+pub use search::{search, search_any, search_one};
